Refactor future booking checks to be synchronous
Replaced asynchronous `CheckFutureBookingsAsync` with synchronous `CheckFutureBookings` across the codebase. Updated the `DeletePrivateCustomerAsync` method to handle the synchronous behavior and added a conditional check to throw a `PrivateCustomerException` if future bookings exist. Modified the `IPCustomerFutureBookingChecker` interface and its implementation to reflect this change. Improved error handling and user feedback for deleting customers with future bookings.

+++ BellaHair.Application/PrivateCustomers/PrivateCustomerCommandHandler.cs
@@ -62,7 +62,8 @@ namespace BellaHair.Application.PrivateCustomers
         {
             var customerToDelete = await _privateCustomerRepo.GetAsync(command.Id);
 
-            _pCustomerFutureBookingChecker.CheckFutureBookingsAsync(customerToDelete);
+            if (_pCustomerFutureBookingChecker.CheckFutureBookings(customerToDelete)) 
+                throw new PrivateCustomerException("Customer has future bookings. Delete bookings prior to deleting customer.");
 
             _privateCustomerRepo.Delete(customerToDelete);
